Reorganize issue alert files
