+++ src/sentry/integrations/discord/actions/__init__.py
@@ -1,2 +1,2 @@
-from .form import *  # noqa: F401, F403
-from .notification import *  # noqa: F401, F403
+from .issue_alert.form import *  # noqa: F401, F403
+from .issue_alert.notification import *  # noqa: F401, F403
