+++ src/sentry/integrations/discord/actions/issue_alert/notification.py
@@ -2,7 +2,7 @@ from typing import Any, Generator, Sequence
 
 from sentry import analytics, features
 from sentry.eventstore.models import GroupEvent
-from sentry.integrations.discord.actions.form import DiscordNotifyServiceForm
+from sentry.integrations.discord.actions.issue_alert.form import DiscordNotifyServiceForm
 from sentry.integrations.discord.client import DiscordClient
 from sentry.integrations.discord.message_builder.issues import DiscordIssuesMessageBuilder
 from sentry.rules.actions import IntegrationEventAction
